% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_sim}
\alias{dat_sim}
\title{Simulated Data with a Partial Ordering}
\format{
A data frame with 1000 observations and 9 numerical variables simulated by
drawing from a multivariate distribution according to the DAG \code{true_sim}.
\describe{
  \item{A1}{numeric}
  \item{B1}{numeric}
  \item{C1}{numeric}
  \item{A2}{numeric}
  \item{B2}{numeric}
  \item{C2}{numeric}
  \item{A3}{numeric}
  \item{B3}{numeric}
  \item{C3}{numeric}
}
}
\usage{
dat_sim
}
\description{
A simple graph and corresponding dataset used in the examples illustrating \code{\link{tpc}}.
}
\keyword{datasets}
