% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpc-package.R
\docType{package}
\name{tpc-package}
\alias{tpc-package}
\alias{_PACKAGE}
\title{Temporal PC Algorithm}
\description{
Constraint-based causal discovery using the PC algorithm while accounting for a
partial node ordering, e.g. a partial temporal ordering when the data were
collected in different waves of a cohort study.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bips-hb/tpc}
  \item Report bugs at \url{https://github.com/bips-hb/tpc/issues}
}

}
\author{
\strong{Maintainer}: Ronja Foraita \email{foraita@leibniz-bips.de} [contributor]

Authors:
\itemize{
  \item Janine Witte \email{witte@leibniz-bips.de}
}

Other contributors:
\itemize{
  \item DFG [funder]
}

}
\keyword{internal}
