% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{true_cohort}
\alias{true_cohort}
\title{Cohort Data Structure}
\format{
A DAG (graphNEL object) with 34 nodes and 128 edges.
}
\usage{
true_cohort
}
\description{
A DAG from which the data 'data_cohort' was simulated from.
See Andrews et al. (2021) <https://arxiv.org/abs/2108.13395> for more information on
how the data were generated.
}
\references{
Andrews RM, Foraita R, Witte J (2021). A practical guide to causal discovery
with cohort data. <https://doi.org/10.48550/arXiv.2108.13395>
}
\seealso{
See [graph::graphNEL()] for the class 'graphNEL'.
}
\keyword{datasets}
