% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{true_sim}
\alias{true_sim}
\title{A DAG with a Partial Ordering}
\format{
A DAG (graphNEL object) with 9 nodes and 7 edges.
}
\usage{
true_sim
}
\description{
An example DAG from which the data 'data_sim' was simulated from.
}
\seealso{
See [graph::graphNEL()] for the class 'graphNEL'.
}
\keyword{datasets}
